More pedantic fixes Fixed makefile for CI

+++ source/utilities/time.cpp
@@ -11,7 +11,7 @@ namespace utilities
     double time::get_current_time_in_seconds()
     {
         const auto time_now_in_nano_seconds = std::chrono::high_resolution_clock::now().time_since_epoch();
-        return time_now_in_nano_seconds.count() / NANOSECONDS_IN_ONE_SECOND;
+        return static_cast<double>(time_now_in_nano_seconds.count()) / static_cast<double>(NANOSECONDS_IN_ONE_SECOND);
     }
 
     void time::sleep_sec(uint32_t seconds)
